+++ packages/core/src/utils/formatQuery/index.ts
@@ -60,6 +60,7 @@ export * from './defaultRuleGroupProcessorPrisma';
 export * from './defaultRuleGroupProcessorSequelize';
 export * from './defaultRuleGroupProcessorSpEL';
 export * from './defaultRuleGroupProcessorSQL';
+export * from './defaultRuleGroupProcessorValidation';
 export * from './defaultRuleProcessorDrizzle';
 export * from './defaultRuleProcessorElasticSearch';
 export * from './defaultRuleProcessorJSONata';